Deduplicate Store data into database
